Fix melon not works on newest and old versions
There was a problem when using the melon_block material on the older version (1.12) for the Break action, it did not detect even if a simple melon was specified. But on the other hand, it worked on the latest (1.15) versions.

So now the correct way is using the legacy_melon_block when using oldest mc versions.