Deprecated short description in jobs now we should use FullDescription
Now the FullDescription list will be printed to the gui items instead of short description. In fact, some type didn’t make sense, as even simple strings could solve new lines, but many were lazy to do it, with just two characters. So the FullDescription list will be visible on new servers that have just installed jobs or are updating the file manually.

- Added potion effect support to jobs shop to give potion item to player with the given effect type.

Closes #1064