feat(tops): add possibility to be hidder from tops
To do that, the part of the command 'top' and 'gtop' has been moved into
an asynchronous thread because of 'unsafe Vault lookups from LuckPerms'.
If it enable, LP prevent a plugin to check a perm of an online player on
the main thread.

Two new nodes perms have been added :
- `jobs.hidetop.<job>`: to be hidden from a specific job top (replace
  <job> by * to be hidden from all).
- `jobs.hidegtop`: to be hidden from the jobs global top.